More error cleanups.
Handle the error codes from rte_eal_init().  Treat EACCES as a
permissions error when activating devices.  Treat ENOTSUP as meaning
"DPDK isn't available on this machine", which isn't a fatal error for
pcap_findalldevs() - we just don't find any DPDK devices.  Treate
EALREADY as meaning "we already called this", and note that we've done
so.  For other errors mentioned in the DPDK documentation, provide error
message strings as per the documentation for the error.

For pcap_findalldevs(), if we have no available Ethernet ports, just
report no devices, don't report a fatal error.