Clean up string allocation.
Don't format into an on-the-stack buffer and copy it to a mallocated buffer.

Instead, mallocate the buffer for the name first, and then format into
*that* buffer.

This avoids the risk of truncating the string when formatted into the
on-the-stack buffer, squelching some compiler warnings and possibly
squelching truncation.

Use strdup() to mallocate a copy of a string - don't reimplement
strdup() in a more complicated fashion, just use it.

(cherry picked from commit b8788a74caa124432f92f48a19651d10f0f18b30)