Try building with Xenial as well as Trusty, and doing fewer macOS builds.
Try building with both Xenial and Trusty on Ubuntu, so we can try
testing builds with DPDK, which is only available on Xenial.

Don't build with both gcc and clang on macOS for now; according to the
Trusty documentation, gcc is just an alias for clang, and to really
compile with *GCC*, you'd have to set CC in the environment to a
specific GCC version, and you'd have to use "brew install" to get
versions other than 4.9.