OK, try *this*.
The syntax, including the indentation, is taken from a .travis.yml file
for a GitHub repository for a project being maintained by "Twisted
Matrix Labs", who are apparently responsible for some obscure product
called "Twisted". :-)

Presumably *that* works.  The only things they're *not* doing are 1)
tweaking the distribution and compiler and 2) setting multiple
environment variables.  If Travis can't parse *that*, what's the issue?