OK, go back to the explicit matrix, and fix package installation.
It appears that having multiple distributions in dist: causes the build
to hang forever in the queueing stage.  Go back to the explicit matrix
as a way to get both Trusty and Xenial builds.

If you have an item in some matrix entries and the same item outside the
matrix, the one in the matrix entry overrides the item outside the
matrix, so we have to have the APT add-on packages in each matrix entry.
Do so.