Use asprintf(), and provide a version for systems that lack it.
This lets us get rid of places where we calculate the length that a
formatted string will require, attempt to allocate a buffer for the
string, and then format the string into that buffer; this way, we don't
have to hope we got the calculation correct, we rely on the same code
that formats strings to do the calculation.

Provide versions of asprintf() for:

1) Windows and the MSVC runtime, where we use _vscprintf() to determine
the length;

2) systems that have (what is presumed to be) an snprintf() that, when
handed a too-short buffer, returns the number of characters that would
have been written had the buffer been long enough (as C99 specifies),
where we format to a one-character buffer to determine the length;

3) systems that don't have snprintf(), where we use the asprintf()
provided by the missing/snprintf.c file that also provides snprintf().

While we're at it, include "portability.h" in missing/win_snprintf.c, to
get declaration/definition checking done.