Forget jessie-backports.
Either it or Travis or both appear to require a GPG key to verify the
packages, and Debian appears not to make it easy to determine what the
URL is to get it (or don't make it available via HTTP).

Thanks, fellas.  I really appreciate how easy you made it to test
building libpcap with a sufficently-modern DPDK on Travis.