Don't treat EACCES as a fatal error when enumerating DPDK devices.
Unfortunately, pcap_findalldevs() has no mechanism to return warnings,
and a permissions error in a particular module's find-devices routine
ultimately be a warning - only if *all* modules fail with "permission
denied" should we treat it as a fatal error.

So, if we're just trying to enumerate devices, treat EACCES as meaning
DPDK is unavailable, not as a fatal error.