Get rid of GNU Makeisms.
We don't (and shouldn't) require GNU Make; instead of using ifeq(), have
PLATFORM_C_SRC and PLATFORM_CXX_SRC variables, set in the configure
script, that contain the platform-specific C and C++ source files.

While we're at it, have MODULE_C_SRC for add-on module source and
REMOTE_C_SRC for remote-capture source, and rename CSRC to COMMON_C_SRC
and GENSRC to GENERATED_C_SRC.

Note that, as not all source files are necessarily C files, we have to
expand each component of OBJ separately, at least to the extent of
expanding _C_SRC and _CXX_SRC variables separately.