Rename DLT_IPMB to DLT_IPMB_KONTRON.
Apparently the Kontron person who said that the packets begin with the
I2C slave address didn't understand the question, because the Kontron
code adds a 2-byte pseudo-header before the I2C slave address.  Rename
the DLT_ and LINKTYPE_ values to reflect this, and update the comments.

Add support for the Linux/Pigeon Point pseudo-header in some places
where it was missing.