Always define pcap_EXPORTS even if we're not calling the library target "pcap".
That's what the header files check for, so we need to define
pcap_EXPORTS no matter what we're calling the library, so that, for
example, if it's being built as "wpcap" - which might be the case on
Windows, for WinPcap compatibility - we get the symbols from pcap/pcap.h
etc. exported rather than imported.

Should address GitHub issue #810.

(cherry picked from commit 4d490932e08daefa778f58ea4694fc97486507ce)