Use "ab+" rather than "rb+" for the pcap_dump_open_append() open mode.
That way, the file will be created if it doesn't exist.

Do a seek to the beginning of the file before checking whether it has a
file header; ISO C says it's implementation-defined whether an open in
append mode puts the file position indicator at the beginning or the end
of the file, and neither the Single UNIX Specification nor the Microsoft
documentation seemed to indicate which of those are the case.

Consistently cast the result of fclose() on an open failure to void -
we've already gotten an error, so we don't care whether the close fails
as well.

Clean up an error message slightly.

See GitHub issue #247.