Test hdr.snaplen to see whether it fits in an int.
Assigning it to p->snapshot, and then checking whether the result is
negative, should work in practice, but it gets unsigned-behavior
warnings.  Test beforehand whether it's valid, and only assign it to
p->snapshot if it is.

This should address the pcap.c part of GitHub issue
the-tcpdump-group/tcpdump#785.

(cherry picked from commit dd0edaf7eb97c05aff6502056c1e5de9944eb209)