Remove some workarounds for old compilers.
Require Visual Studio 2015 or later; fail if we don't have it, and
remove checks for older versions.

That means we have C99-compliant snprintf() and vsnprintf(); require
them when configuring for UN*X, and then use them directly, rather than
having wrappers for systems lacking them.

If we're using MSVC, skip the tests for options to request C99
compatibility - either we have VS 2015, which is sufficient, or we
don't, in which case we fail.