Get the ARPHRD_ values from a Linux kernel header.
linux/if_arp.h is a "uapi" header, so it's part of the user-mode API to
the Linux kernel, and we're user-mode code that's using Linux-specific
kernel APIs, so it's what we should be using - especially given that C
libraries that provide a <net/if_arp.h> header don't necessarily keep
them up-to-date with all the kernel ARPHRD_ values.

(I.e., skip the middleman; the kernel is the appropriate authority as to
what ARPHRD_ values it returns.)

This should address the problem in GitHub pull request #836.