Always return a list of supported time-stamp types.
In change 6e0c54a8f9640973e95ae755e6f60e1cceb7985f, we "have both the
time stamp type-setting and precision-setting routines allow the default
value (PCAP_TSTAMP_NONE/PCAP_TSTAMP_PRECISION_MICRO) if no list is
supplied."  That means that, for example, "tcpdump -J" may report "Time
stamp type cannot be set", but "tcpdump -j host" is allowed.  One could
argue that you're not "setting" the time stamp, as "host" is the
default, but it's confusing.

This means that "host" will be reported as the one and only time stamp
type if that's supported.