Make sure pcap_close_all() doesn't loop infinitely.
After it calls pcap_close(), the pcap_t that was closed must no longer
be at the head of the pcaps_to_close list.  Abort if it still is, rather
than looping indefinitely (or crashing because we're trying to close a
now-closed-and-freed pcap_t).

Reported by Charles Smith at Tangible Security.