Try to handle URL paragraphs better.
We'd like this document to show links as links when formatted as
Markdown *and* to have it be at least somewhat readable as plain text,
so we want to have the URLs set off somewhat for the latter.  Try block
quoting.

(cherry picked from commit 910c67a72abc5792e2088e5c691ca17ee91ba0a3)