More pcap_setdirection() cleanups.
Say "...on this device" for errors; it may be the OS's capture mechanism
that doesn't support setting the direction, but there may be other
devices on the system that *do* support it.

Check the validity of the direction argument in more places.

(backported from commit 2c2c323c392980edc4a1044f08d9c2889f336dd3)