We don't check the direction, so don't support setting it.
If there's a direction indicator in the packets we get, we should allow
the direction to be set and should check it if it's "incoming only" or
"outgoing only".  Otherwise, we shouldn't allow the direction to be set.

(cherry picked from commit 25a2cbb98b22171312ec894b906d56af6344dfeb)