More cleanups.
Take an else clause for which the if clause returns out and put it at
the top level.

Rename activate_mmap() to setup_mmapped(), as it does part, not all, of
the activation process.  Put a bit more of that process into it, so it
finishes the process (except for the final "set the protocol so packets
are seen" part).

Add a setup_non_mmapped() routine that finishes the non-memory-mapped
setup process (except for the final "set the protocol so packets are
seen" part).

(backported from commit 769533f18b0f855acf6477841cb01f20b8dfd3a8)