Work around for bogus definition of _MSC_VER with non-MS compilers.
Try to work around software using pcap that defines _MSC_VER with non-MS
compilers before including pcap.h.

We expect _MSC_VER to be defined only for Microsoft's compiler, and to
be defined with a value that reflects what version of the compiler is
being used, so that we can determine, among other things, what Microsoft
extensions are available.

Try to detect it being defined by other software that wants to trick us
into thinking something's being compiled with Microsoft's compiler, and
undo the definition of _MSC_VER.