Clean up search for OpenSSL/libressl.
Don't just throw in -I/usr/local/opt/openssl/include and
-L/usr/local/opt/openssl/lib if we're on macOS and have found
OpenSSL/libressl; that may not be where it is.

Instead, if we don't find it in the "usual" directories or with
pkg-config, explicitly check for a usable OpenSSL/libressl under
/usr/local/opt/openssl and, if we find it, set up OPENSSL_CFLAGS and
OPENSSL_LIBS the same way we do it elsewhere.