Don't test whether we have NdisMediumWirelessWan with an #ifdef.
1) As noted by Daniel Miller in nmap/nmap#1573, NdisMediumWirelessWan is
a member of an enum, not a #define, so you can't test its presence with

2) It should be present in Windows 7 or later SDKs; if that means that
any sufficiently recent build environment should have it, it might not
be worth worrying about older build environments, especially given that
we require a bunch of C99isms to be supported, so you'll need a newer
Visual Studio version anyway.

If it turns out to be an issue, we could add CMake checks for
NdisMediumWirelessWan.