Check whether a name is too longer before using it in an ioctl.
If the name is longer than the name field in the structure passed in to
the ioctls, it's longer than is valid for an interface name, so it
definitely doesn't exist.

Do length checking in bpf_bind() and device_exists().

Have bpf_bind() return a PCAP_ERROR_ code, doing the error
interpretation work previously done by its callers.  Have it return a
special code for "buffer too big", so that code that loops, reducing the
buffer size, trying to pick a default buffer size knows when to continue
looping.