autotools: add the DAG include dir to CFLAGS before looking for the header.
And don't set V_INCLS unless we find it.

While we're at it, clean up the save/restore of linker flags in various
places to 1) use a variable with a name that matches the style used
elsewhere (save_LDFLAGS rather than saved_ldflags) and 2) quote the
flags, in case there are blanks in the string.

This should fix GitHub issue #932.