dpdk: check for rte_eth_dev_count_avail().
We call it, but older versions of DPDK didn't have it, so check for it.

Fix the autotools error messages issued if we find DPDK but it lacks
rte_bus.h or rte_eth_dev_count_avail() to indicate that the problem is
that the DPDK is too old.

Use AC_CHECK_HEADER(), not AC_CHECK_HEADERS(), to check for rte_bus.h;
we don't need a #define to indicate whether we have it, we just refuse
to build DPDK support if we don't have it.