SSL: attempt to squelch a narrowing warning.
SSL_set_fd() takes an int as its second argument; that's not an issue on
UN*X, as sockets are represented by file descriptors, but causes
warnings on Windows, where a SOCKET is not an int.

A comment in OpenSSL's intternal sockets.h header says, on 64-bit
Windows:

	Even though sizeof(SOCKET) is 8, it's safe to cast it to int,
	because the value constitutes an index in per-process table of
	limited size and not a real pointer.

so the cast should be safe.