Add support for setting the snapshot length and not setting the filter.
Add a -s option to set the snapshot length; don't set it, even to a
default value, if -s isn't specified.

If there are no arguments specifying a filter, don't compile or install
a filter.

This allows us to test 1) whether the snapshot length works even if no
filter is set and 2) what happens if the snapshot length isn't set.