NPF: squelch some warnings about enum values not handled in switches.
The AppVeyor CI builder started warning about switch statements in which
not all enum values were handled explicitly on the VS 1019 build.  We
can't test for the enum values at compile time, the way we could test
for #defines, so, for now, we punt on the NdisPhysicalMedium enums and
just suppress that diagnostic for that switch statement.

We add MediaConnectStateUnknown to the switch for the connect state;
that one is presumably there if the enum is there at all.