configure: don't look for libnl in a .libs directory.
As I remember, that's some sort of libtool thing, which shows up in
build directories but not in install directories, so that would mean
--with-libnl=XXX would work if you were pointing it at a build directory
(*if* libnl uses libtool) but wouldn't work if you were pointing it at
the tree under which libnl was installed.

This should address the main part of GitHub issue #952.