doc: give more painful details about packet time stamps.
Mention the missing leap seconds problem (which you will see on UN*Xes
if the clock keeps POSIX time, and which you might see on Windows as
well).

Note that the big list of problems means that subtracting time stamps
might not give a precise value for the interval between packets.

Note that device time stamps synchronized with host clocks may have
some of the problems that host time stamps have.

[skip ci]