Fixup rpcapd destination in CMake. [skip ci]
Autoconf installs rpcapd into sbindir, which is consistent with the
logic of recent changes in tcpslice and tcpdump: rpcapd is useless to
non-root users unless the distribution provides custom means to elevate
the privileges (but in that case installing into a custom directory
would be the least problem). CMake, however, for some reason has been
installing into bindir, so make it consistent with autoconf, at least on
non-Windows systems.