Run Bison with -Wno-yacc.
Autoconf's AC_PROG_YACC turns on the -y flag, which, for Bison 2.5 and
later, causes it to print warnings about features not supported by POSIX
YACC; we don't care, we only care about Bison and Berkeley YACC, so we
suppress that warning with -Wno-yacc.

While we're at it, we already determine whether we're running Bison or
Berkeley YACC before we create the custom command for the scanner, so,
for CMake, add the -y flag there.

(cherry picked from commit 19fa79eb9842d1e6cdd63718f23c9d60bb5c1eaf)