Do our own search for Bison/Berkeley YACC.
If AC_PROG_YACC finds Bison, it arranges to run it with -y, for YACC
compatibility, but we don't care about YACC compatibility, and, with
newer versions of Bison, -y also causes annoying warnings about our
grammar using features not in POSIX YACC, which is another thing about
which we don't care - we don't support POSIX YACC, we only support
versions of Bison and Berkeley YACC that support reentrant parsers.

This means we don't need to turn those warnings off with -Wno-yacc.