Refine Markdown in README.md. [skip appveyor]
Move "by TTG" to the main header. Switch from the alternative headings
format, which allows for H1 and H2 only, to the main one, which allows
for H1 through H6. Have two H3 headers in the text. Fix CI badges to
point to libpcap instead of tcpdump, and to use .com URLs for Travis CI.
Make the security note bold to help more people actually get it. Add and
improve hyperlink formatting in a few places. Fence the "from LBL"
boilerplate and move it to the bottom of the first section. Make the
"NOT", which was bold in commit 4a0b801, but didn't survive the initial
Markdownization attempt in commit 82001f3, bold again. Escape filenames
and sonames properly.

Skip only some CI on this occasion, so Travis CI can transfer the build
history from travis-ci.org to travus-ci.com and complete the migration.