Drop support for text-mode Linux USB captures.
We require a 2.6.27 or later kernel, so, on the platforms that we
support, binary-mode captures are supported.

This gets rid of some sscanf() calls that could overflow string buffers
if the kernel has malicious code in the support for text-mode USB
captures.

(cherry picked from commit a9e9bca9806941662caa0ebc0e354bbb913296d4)