Fix Bison detection for minor version 0. [skip ci]
Amend the autoconf and CMake regexps to match a minor version that is
an integer >= 0 instead of >= 1. As far as I can tell, grammar.y would
have "%define api.pure" instead of "%pure-parser" for bison/yacc
versions 1.0 and 2.0 (cannot estimate the probability of running into
these in the wild though). Also this bug caused an error message in the
middle of the usual output:

./configure: line 8569: test: : integer expression expected

$ bison -V
bison (GNU Bison) 3.0.4
Written by Robert Corbett and Richard Stallman.

Copyright (C) 2015 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.