Bluetooth: fix non-blocking mode
Fix non-blocking mode with bt-linux and bt-monitor-linux. Previously
when fd was set to non-blocking mode and pcap_dispatch() was called
from application, it returned -1 with errno EAGAIN breaking the api
contract.

Fix is to handle errnos EAGAIN and EWOULDBLOCK so that 0 is returned.

(cherry picked from commit 747ca9149a4158fafc4551550976cdb028916182)