CI: Refine the build matrix scripts. [skip appveyor]
If not provided with a prefix path, make a temporary one and remove it
when done; between the rounds delete the contents of the prefix
directory, but not the directory. This way parallel builds on a buildbot
worker do not clash, also if a developer runs build_matrix.sh from their
home directory, that no longer breaks the worker builds whilst the
non-worker script is running or even indefinitely if it had failed.

Fix skipping GCC on macOS in build_matrix.sh to make it work
consistently in and outside of a CI environment. Remove the skipping
from the CI space.

Quote the prefix variable properly. Print OS and compiler versions at
the beginning. Simplify Makefile printing in the Travis CI block.
Require travis_fold() parameters to be set. Relabel Cirrus CI tasks for
consistency.

(cherry picked from commit 765350472ba83842ece8d2f6aaa70993bda99fa6)