CMake: fix up generation of libpcap.pc when not building a shared library.
Don't unset LIBRARY_NAME if we're not building shared libraries; that's
confusing and weird.  Instead, set a variable LIBRARIES_TO_INSTALL to
the list of libraries to install, so that it includes the shared library
if we build it and doesn't include the shared library if we don't.  Use
that when setting up installation.

Set PACKAGE_NAME only once, as we're not changing LIBRARY_NAME.

The bug was that

1) if we aren't building shared libraries, we were clearing LIBRARY_NAME

and

2) we set PACKAGE_NAME twice - once before LIBRARY_NAME was cleared and
once after it was cleared in the non-shared-libraries case, and we used
the *latter* version to generate libpcap.pc

so it would have an empty library name in the no-shared-libraries case.

Fixes GitHub issue #1009.