npf: group the NdisMedium values into two sorted groups.
The first group is the group of NdisMedium values defined by Microsoft
as part of the enum; the second gropu is the group of values defined by
Npcap for its own purposes.  Both are sorted numerically, to make it
easier to determine which ones we have DLT_ values to which to map them.