Don't require VS 2015 or later when building applications with pcap.
Some Npcap customers are using older versions of VS, so define the
relevant integer types ourselves in that case, so that the headers we
provide that use them can be used.

However, when building libpcap *itself*, require VS 2015 or later, as we
assume a level of C99 support found only in VS 2015 or later.

(cherry picked from commit 17f04d83389cb3698a043bf968c8651d7776ceee)