rpcapd: note the alignment issue, get rid of an extra pointer variable.
Add a comment noting why we aren't just filling in structures in the
output buffer when constructing a findalldevs response.

Instead of naming the address structures "tmp", and having a pointer to
them, just give them regular names and fill them in without a pointer.

Make some white space stylistic changes while we're at it.