rpcap: don't do pointless integer->string and then string->integer conversions.
The string->integer conversion was also broken, as it passed a pointer
to a 16-bit integer to a sscanf() call that used %d rather than %hd.
It'd overwrite 2 bytes past the 16-bit integer; it may set the integer
"correctly" on a little-endian, but wouldn't even do *that* on a
big-endian machine.

(cherry picked from commit efaddfe8eae4dab252bb2d35e004a40e4b72db24)